+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # lumbermixalot Changelog
 
+## v2.0.0
+Made changes to make Actor & Motions compatible with O3DE (https://github.com/o3de/o3de)
+Please use v1.1.1 if you are still working with Lumberyard 1.2X.
+
 ## v1.1.1
 Minor bug fix. changed **description** to **bl_description** for bpy.types.Operator subclasses in __init__.py
 Reason: Blender 2.81 was not lenient anymore when misnaming bl_description.